Update dependency on puppetlabs/apt
Breaking changes for puppetlabs/apt 5.0.0 was a deprecation of Debian 7;
for 6.0.0, it was removal of support for Puppet <= 4.7.

Since the bareos module itself is not affected by syntax changes, we can
leave the dependency on apt pretty wide so users can choose a version of
apt supported on their OS version.