Allow multiple occurences of sub-options in filesets
Bareos filesets support definitions like

```
Include = {
  Options = { ...
  }
  Options = { ...
  }
  ...
}
```

So far, there was only slightly hacky support for doing this with the
arrays of strings inside, but now there's an even hackier approach for
this also allowing arrays hashes!

This whole thing needs a rewrite, though, as almost all of the
validation can be outsourced to type aliases. This will follow in a
later PR.